Move DynamicXPLayer to contrib/common/ofa/elastic_nn/modules/dynamic_layers.py
